// oline_fish_shope_frontend/src/app/enums/category.enum.ts
export enum Category {
  Anchovies = "Anchovies",
  ArcticChar = "Arctic Char",
  BaconWrapped = "Bacon Wrapped",
  BassBranzino = "Bass/Branzino",
  BliniCremeFraiche = "Blini/Crème Fraîche",
  Books = "Books",
  Bundles = "Bundles",
  Burgers = "Burgers",
  Cakes = "Cakes",
  CannedSeafood = "Canned Seafood",
  Caviar = "Caviar",
  Clams = "Clams",
  Clothing = "Clothing",
  Cod = "Cod",
  Crab = "Crab",
  CuratedByFulton = "Curated By Fulton",
  FeaturedSeafood = "Featured Seafood",
  Fillets = "Fillets",
  Halibut = "Halibut",
  KitchenTools = "Kitchen Tools",
  LiveFreshShellfish = "Live/Fresh Shellfish",
  Lobster = "Lobster",
  LobsterMeat = "Lobster Meat",
  Mussels = "Mussels",
  Octopus = "Octopus",
  Oysters = "Oysters",
  Porgy = "Porgy",
  PortionsAndSteaks = "Portions & Steaks",
  Salmon = "Salmon",
  SaucesSpices = "Sauces/Spices",
  Scallops = "Scallops",
  Shrimp = "Shrimp",
  SmokedSeafood = "Smoked Seafood",
  Snapper = "Snapper",
  Squid = "Squid",
  Swordfish = "Swordfish",
  Trout = "Trout",
  Tuna = "Tuna",
  Uni = "Uni (Sea Urchin)",
  Whiting = "Whiting",
  WholeFish = "Whole Fish"
}