// oline_fish_shope_frontend/src/app/enums/department.ts
export enum Department {
  CAVIAR_ROE = "Caviar & Roe",
  CLAMS_OYSTERS = "Clams / Oysters",
  CRAB_LOBSTER_SHRIMP = "Crab / Lobster / Shrimp",
  CURATED_BY_FULTON = "Curated By Fulton",
  FISH = "Fish",
  MUSSELS_SCALLOPS = "Mussels / Scallops",
  OCTOPUS_SQUID = "Octopus / Squid",
  PANTRY_MERCHANDISE = "Pantry / Merchandise",
  PREPARED_READY_TO_EAT = "Prepared / Ready To Eat",
}